decode/pyproject.toml
[project]
name = "2023-05-07457A"
version = "1.0.0"
description = "Decoding analysis for Distinct brain-wide presynaptic networks underlie the functional identity of individual cortical neurons"
authors = [
    { name = "yuanz", email = "yuan.zhao@nih.gov" }
]
dependencies = [
    "numpy>=1.24.4",
    "scipy>=1.10.1",
    "matplotlib>=3.7.3",
    "seaborn>=0.12.2",
    "pandas>=2.0.3",
    "polars>=0.19.3",
    "scikit-learn>=1.3.0",
    "pymatreader>=0.0.32",
    "pyyaml>=6.0.1",
    "jupyterlab>=4.1.6",
    "absl-py>=2.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
